+++ src/main/java/com/meli/desafio_spring/users/models/User.java
@@ -1,19 +1,19 @@
 package com.meli.desafio_spring.users.models;
 
 public abstract class User {
-    private long userId;
+    private int userId;
     private String userName;
 
-    public User(long userId, String userName) {
+    public User(int userId, String userName) {
         this.userId = userId;
         this.userName = userName;
     }
 
-    public long getUserId() {
+    public int getUserId() {
         return userId;
     }
 
-    public void setUserId(long userId) {
+    public void setUserId(int userId) {
         this.userId = userId;
     }
