+++ src/main/java/com/meli/desafio_spring/users/repository/imple/DatabaseMethods.java
@@ -1,6 +1,6 @@
 package com.meli.desafio_spring.users.repository.imple;
 
 public interface DatabaseMethods<T> {
-    T findById(Long id);
+    T findById(int id);
     void save(T item);
 }
