+++ src/main/java/com/meli/desafio_spring/users/service/UserService.java
@@ -20,7 +20,7 @@ public class UserService {
     @Autowired
     private SellerRepository sellerRepository;
 
-    public void followSeller(Long buyerId, Long sellerId){
+    public void followSeller(int buyerId, int sellerId){
         Seller seller = sellerRepository.findById(sellerId);
         Buyer buyer = buyerRepository.findById(buyerId);
         if(seller == null || buyer == null)
@@ -33,25 +33,25 @@ public class UserService {
         buyerRepository.save(buyer);
     }
 
-    public Seller getSeller(Long userId) {
+    public Seller getSeller(int userId) {
         Seller seller = sellerRepository.findById(userId);
         if(seller == null)
             throw new ObjectNotFoundException();
         return seller;
     }
 
-    public Buyer getBuyer(Long userId) {
+    public Buyer getBuyer(int userId) {
         Buyer getBuyer = buyerRepository.findById(userId);
         if(getBuyer == null)
             throw new ObjectNotFoundException();
         return getBuyer;
     }
 
-    public List<Buyer> getBuyers(List<Long> buyersIds){
+    public List<Buyer> getBuyers(List<Integer> buyersIds){
         return buyersIds.stream().map(id -> buyerRepository.findById(id)).collect(Collectors.toList());
     }
 
-    public List<Seller> getSellers(List<Long> sellersIds){
+    public List<Seller> getSellers(List<Integer> sellersIds){
         return sellersIds.stream().map(id -> sellerRepository.findById(id)).collect(Collectors.toList());
     }
 }
